requirements.txt
rio-ui
newsapi-python==0.2.6
openai
peewee